sql-agent-openai/requirements.txt
chainlit>=2.8.4
openai>=1.12.0
pyodbc>=5.0.1
mcp>=1.19.0
python-dotenv>=1.0.0
httpx>=0.26.0
fastapi>=0.109.0
uvicorn>=0.27.0





